Replace Floyd-Warshall with simpler approach
Since the only possible distance between two neighboring cells is 1 and there is no such thing as unreachable node then there is no need in Floyd-Warshall. We know the shape of the grid and we can just calculate position of each node. Then we count number of steps needed to get to that node from each other node. This reduces complexity from O(n^3) to O(n^2). In my case of SOM with side 25 (625 nodes) that was an improvement from 4 minutes to 18 seconds (total algorithm time).